limiting threads

+++ android/src/main/java/com/GetMusicFiles/Utils/FS.java
@@ -2,6 +2,11 @@ package com.GetMusicFiles.Utils;
 
 import android.util.Log;
 
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.Formatter;
+
 import androidx.annotation.NonNull;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -13,9 +18,27 @@ import static com.GetMusicFiles.Utils.GeneralUtils.LOG;
 
 public class FS {
 
+    private static String SHAsum(byte[] byteArray){
+        MessageDigest md = null;
+        try {
+            md = MessageDigest.getInstance("MD5");
+        } catch (NoSuchAlgorithmException e) {
+            e.printStackTrace();
+        }
+        return byteArray2Hex(md.digest(byteArray));
+    }
+
+    private static String byteArray2Hex(final byte[] hash) {
+        Formatter formatter = new Formatter();
+        for (byte b : hash) {
+            formatter.format("%02x", b);
+        }
+        return formatter.toString();
+    }
+
     public static String saveToStorage(@NonNull String pathToImg, @NonNull byte[] imageBytes) throws IOException {
         File f = new File(pathToImg);
-        File filePath = new File(pathToImg+ "/" + UUID.randomUUID().toString());
+        File filePath = new File(pathToImg+ "/" + SHAsum(imageBytes));
 
         if(f.exists() && !f.isDirectory()){
             Log.e(LOG, "coverPath is a file");
@@ -26,10 +49,16 @@ public class FS {
             Log.d(LOG, "coverFolder created");
         }
 
+        if(filePath.exists()){
+           return filePath.getAbsolutePath();
+        }
+
         FileOutputStream fos = null;
         try {
             fos = new FileOutputStream(filePath, false);
             fos.write(imageBytes);
+            fos.flush();
+            fos.close();
         } catch (IOException e) {
             if (fos != null) {
                 fos.flush();
@@ -43,7 +72,7 @@ public class FS {
                 return  filePath.getAbsolutePath();
             }
         }
-        return "";
+        return "-1";
     }
 
     public static void listFilesForFolder(File folder, int minFileSize, int maxFileSize, String extensionFilter, List<String> list) {
